fix two typing errors in TCK
1. the type of the count() function is Long in JPA
   (we should clarify this in JDQL b/c it is undefined)
2. we say that only things with the exact same type
   can be compared using =, <>, and friends

+++ tck/src/main/java/ee/jakarta/tck/data/framework/read/only/PositiveIntegers.java
@@ -70,7 +70,7 @@ public interface PositiveIntegers extends BasicRepository<NaturalNumber, Long> {
 
     // Per the spec: The 'and' operator has higher precedence than 'or'.
     @Query("WHERE numBitsRequired = :bits OR numType = :type AND id < :xmax")
-    CursoredPage<NaturalNumber> withBitCountOrOfTypeAndBelow(@Param("bits") int bitsRequired,
+    CursoredPage<NaturalNumber> withBitCountOrOfTypeAndBelow(@Param("bits") short bitsRequired,
                                                              @Param("type") NumberType numberType,
                                                              @Param("xmax") long exclusiveMax,
                                                              PageRequest<?> pageRequest);
